backend/requirements.txt
fastapi
uvicorn
requests
google-generativeai>=0.7.0
python-multipart
python-dotenv
